Test both pipelines
